Apply codestyle AccountMS
