+++ .idea/compiler.xml
@@ -7,8 +7,8 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="AccountMs" />
         <module name="TransactionMs" />
+        <module name="accountms" />
         <module name="ServiceReg" />
         <module name="CustomerMs" />
         <module name="gateway" />
@@ -17,10 +17,10 @@
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="AccountMs" options="-parameters" />
       <module name="CustomerMs" options="-parameters" />
       <module name="ServiceReg" options="-parameters" />
       <module name="TransactionMs" options="-parameters" />
+      <module name="accountms" options="-parameters" />
       <module name="gateway" options="-parameters" />
     </option>
   </component>
