+++ .idea/encodings.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/AccountMs/AccountMs/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/AccountMs/AccountMs/target/generated-sources/openapi/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/AccountMs/accountms/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/AccountMs/accountms/target/generated-sources/openapi/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/CustomerMs/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/CustomerMs/target/generated-sources/openapi/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/Gateway/src/main/java" charset="UTF-8" />
