--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/AccountMsApplication.java
@@ -1,15 +1,0 @@
-package com.danny.AccountMs;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients
-public class AccountMsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AccountMsApplication.class, args);
-    }
-
-}
