--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/clients/CustomerClient.java
@@ -1,25 +1,0 @@
-package com.danny.AccountMs.clients;
-
-import com.danny.AccountMs.exception.BadPetitionException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.HttpClientErrorException;
-
-import java.util.UUID;
-
-@Configuration
-public class CustomerClient {
-
-    @Autowired
-    private FeignCustomerClient customerFeignClient;
-
-    public void validateCustomerId(UUID id) {
-        try {
-            this.customerFeignClient.getCustomer(id);
-        } catch (HttpClientErrorException e) {
-            throw new BadPetitionException("El cliente no fue encontrado en el servicio de cuentas");
-        } catch (Exception e) {
-            throw new RuntimeException("Error inesperado: " + e.getMessage());
-        }
-    }
-}
