--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/exception/BadPetitionException.java
@@ -1,14 +1,0 @@
-package com.danny.AccountMs.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class BadPetitionException extends RuntimeException {
-    private final int statusCode;
-
-    public BadPetitionException(String message) {
-        super(message);
-        this.statusCode = HttpStatus.BAD_REQUEST.value();
-    }
-}
