--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/exception/ConflictException.java
@@ -1,14 +1,0 @@
-package com.danny.AccountMs.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class ConflictException extends RuntimeException {
-    private final int statusCode;
-
-    public ConflictException(String message) {
-        super(message);
-        this.statusCode = HttpStatus.CONFLICT.value();
-    }
-}
