--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/exception/NotFoundException.java
@@ -1,14 +1,0 @@
-package com.danny.AccountMs.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class NotFoundException extends RuntimeException {
-    private final int statusCode;
-
-    public NotFoundException(String message) {
-        super(message);
-        this.statusCode = HttpStatus.NOT_FOUND.value();
-    }
-}
