--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/exception/UnAuthorizedException.java
@@ -1,14 +1,0 @@
-package com.danny.AccountMs.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class UnAuthorizedException extends RuntimeException {
-    private final int statusCode;
-
-    public UnAuthorizedException(String message) {
-        super(message);
-        this.statusCode = HttpStatus.UNAUTHORIZED.value();
-    }
-}
