--- AccountMs/AccountMs/src/main/java/com/danny/AccountMs/model/TipoCuenta.java
@@ -1,6 +1,0 @@
-package com.danny.AccountMs.model;
-
-public enum TipoCuenta {
-    AHORRO,
-    CORRIENTE
-}
