--- AccountMs/AccountMs/src/test/java/com/danny/AccountMs/AccountMsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.danny.AccountMs;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AccountMsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
