+++ AccountMs/accountms/src/main/java/com/danny/accountms/AccountMsApplication.java
@@ -1,0 +1,15 @@
+package com.danny.accountms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class AccountMsApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(AccountMsApplication.class, args);
+  }
+
+}
