+++ AccountMs/accountms/src/main/java/com/danny/accountms/exception/BadPetitionException.java
@@ -1,0 +1,15 @@
+package com.danny.accountms.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class BadPetitionException extends RuntimeException {
+
+  private final int statusCode;
+
+  public BadPetitionException(String message) {
+    super(message);
+    this.statusCode = HttpStatus.BAD_REQUEST.value();
+  }
+}
