+++ AccountMs/accountms/src/main/java/com/danny/accountms/exception/ConflictException.java
@@ -1,0 +1,15 @@
+package com.danny.accountms.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class ConflictException extends RuntimeException {
+
+  private final int statusCode;
+
+  public ConflictException(String message) {
+    super(message);
+    this.statusCode = HttpStatus.CONFLICT.value();
+  }
+}
