+++ AccountMs/accountms/src/main/java/com/danny/accountms/exception/NotFoundException.java
@@ -1,0 +1,15 @@
+package com.danny.accountms.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class NotFoundException extends RuntimeException {
+
+  private final int statusCode;
+
+  public NotFoundException(String message) {
+    super(message);
+    this.statusCode = HttpStatus.NOT_FOUND.value();
+  }
+}
