+++ AccountMs/accountms/src/main/java/com/danny/accountms/exception/UnAuthorizedException.java
@@ -1,0 +1,15 @@
+package com.danny.accountms.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class UnAuthorizedException extends RuntimeException {
+
+  private final int statusCode;
+
+  public UnAuthorizedException(String message) {
+    super(message);
+    this.statusCode = HttpStatus.UNAUTHORIZED.value();
+  }
+}
