+++ AccountMs/accountms/src/main/java/com/danny/accountms/model/TipoCuenta.java
@@ -1,0 +1,6 @@
+package com.danny.accountms.model;
+
+public enum TipoCuenta {
+  AHORRO,
+  CORRIENTE
+}
