+++ AccountMs/accountms/src/main/resources/checkstyle-suppressions.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suppressions PUBLIC
+  "-//Checkstyle//DTD Suppressions 1.2//EN"
+  "https://checkstyle.org/dtds/suppressions_1_2.dtd">
+
+<suppressions>
+  <suppress checks="Javadoc" files="."/>
+  <!-- <suppress checks="LITERAL_SWITCH" /> -->
+</suppressions>
+
