+++ AccountMs/accountms/src/test/java/com/danny/accountms/AccountMsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.danny.accountms;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AccountMsApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
